subiendo cambio 2

+++ app/Http/Controllers/BannerController.php
@@ -373,7 +373,6 @@ class BannerController extends Controller
                 $banner->borrado='0';
                 $banner->nivel=$nivel;
 
-
                 if($facultad_id != null && intval($facultad_id) > 0){
                     $banner->facultad_id=$facultad_id;
                 }
@@ -381,12 +380,18 @@ class BannerController extends Controller
                     $banner->programaestudio_id=$programaestudio_id;
                 }
 
-
                 $banner->user_id=Auth::user()->id;
 
-                
-        $band=Banner::where('nivel',$nivel)->where('borrado',0)->where('posision',$posision)->exists();
-
+        if ($nivel==0) {
+            $band=Banner::where('nivel',$nivel)->where('borrado',0)->where('posision',$posision)->exists();
+        }
+        if ($nivel==1) {
+            $band=Banner::where('nivel',$nivel)->where('borrado',0)->where('facultad_id',$facultad_id)->where('posision',$posision)->exists();
+        }
+        if ($nivel==2) {
+            $band=Banner::where('nivel',$nivel)->where('borrado',0)->where('programaestudio_id',$programaestudio_id)->where('posision',$posision)->exists();
+        }
+        
         if ($band==true) {
             $exi='0';
             $msj='El orden de publicación ya existe';       
@@ -446,11 +451,14 @@ class BannerController extends Controller
 
         $result='1';
         $msj='';
+        $exi='';
         $selector='';
 
         $oldImg=$request->oldimg;
 
         $nivel=$request->v1;
+        $facultad_id=$request->v2;
+        $programaestudio_id=$request->v3;
 
         if ($request->hasFile('imagen')) { 
 
@@ -531,8 +539,7 @@ class BannerController extends Controller
             elseif(intval($nivel) == 2){
                 Storage::disk('banerProgramaEstudio')->delete($imagen);
             }
-        }
-        else{
+        }else{
             $input1  = array('posision' => $posision);
             $reglas1 = array('posision' => 'required');
 
@@ -543,53 +550,56 @@ class BannerController extends Controller
                 $result='0';
                 $msj='Debe ingresar la posición del Banner en formato numérico';
                 $selector='txtposisionE';
-            }
-            else{
-
+            }else{
                 if($nombre == null || Strlen($nombre) == 0){
                     $nombre = "";
                 }
 
-                if(strlen($imagen)>0)
-                {
-
-                    if(intval($nivel) == 0){
-                        Storage::disk('banerUNASAM')->delete($oldImg);
-                    }
-                    elseif(intval($nivel) == 1){
-                        Storage::disk('banerFacultad')->delete($oldImg);
-                    }
-                    elseif(intval($nivel) == 2){
-                        Storage::disk('banerProgramaEstudio')->delete($oldImg);
-                    }
-
-
-                    $banner = Banner::findOrFail($id);
-                    $banner->posision=$posision;
-                    $banner->nombre=$nombre;
-                    $banner->url=$imagen;
-                    $banner->activo=$activo;
-                    $banner->user_id=Auth::user()->id;
-
-                    $banner->save();
+                if ($nivel==0) {
+                    $band=Banner::where('nivel',$nivel)->where('borrado',0)->where('posision',$posision)->where('id','<>',$id)->exists();
                 }
-                else
-                {
-                    $banner = Banner::findOrFail($id);
-                    $banner->posision=$posision;
-                    $banner->nombre=$nombre;
-                    $banner->activo=$activo;
-                    $banner->user_id=Auth::user()->id;
-
-                    $banner->save();
+                if ($nivel==1) {
+                    $band=Banner::where('nivel',$nivel)->where('borrado',0)->where('facultad_id',$facultad_id)->where('posision',$posision)->where('id','<>',$id)->exists();
+                }
+                if ($nivel==2) {
+                    $band=Banner::where('nivel',$nivel)->where('borrado',0)->where('programaestudio_id',$programaestudio_id)->where('posision',$posision)->where('id','<>',$id)->exists();
                 }
 
-                $msj='El Banner ha sido modificado con éxito';
-
+                if ($band==true) {
+                    $exi='0';
+                    $msj='El orden de publicación ya existe';       
+                }else{
+                    $exi='1';
+                    if(strlen($imagen)>0){
+                        if(intval($nivel) == 0){
+                            Storage::disk('banerUNASAM')->delete($oldImg);
+                        }
+                        elseif(intval($nivel) == 1){
+                            Storage::disk('banerFacultad')->delete($oldImg);
+                        }
+                        elseif(intval($nivel) == 2){
+                            Storage::disk('banerProgramaEstudio')->delete($oldImg);
+                        }
+                        $banner = Banner::findOrFail($id);
+                        $banner->posision=$posision;
+                        $banner->nombre=$nombre;
+                        $banner->url=$imagen;
+                        $banner->activo=$activo;
+                        $banner->user_id=Auth::user()->id;
+                        $banner->save();
+                    }else{
+                        $banner = Banner::findOrFail($id);
+                        $banner->posision=$posision;
+                        $banner->nombre=$nombre;
+                        $banner->activo=$activo;
+                        $banner->user_id=Auth::user()->id;
+                        $banner->save();
+                    }
+                    $msj='El Banner ha sido modificado con éxito';
+                }
             }
         }
-
-        return response()->json(["result"=>$result,'msj'=>$msj,'selector'=>$selector]);
+        return response()->json(["result"=>$result,'msj'=>$msj,'selector'=>$selector,'exi'=>$exi]);
     }
 
 
@@ -677,13 +687,21 @@ class BannerController extends Controller
 
         return response()->json(["result"=>$result,'msj'=>$msj]);
     }
-    public function numsiguiente($niv)
+    public function numsiguiente($niv,$idfac,$idprog)
     {
         $nivel=$niv;
-        $queryZero=Banner::where('nivel',$nivel)->where('borrado',0)->max('posision');
-
+        $idFacultad=$idfac;
+        $idPrograma=$idprog;
+        if ($nivel==0) {
+            $queryZero=Banner::where('nivel',$nivel)->where('borrado',0)->max('posision');
+        }
+        if ($nivel==1 && $idFacultad!=0) {
+            $queryZero=Banner::where('nivel',$nivel)->where('borrado',0)->where('facultad_id',$idFacultad)->max('posision');
+        }
+        if ($nivel==2 && $idPrograma!=0) {
+            $queryZero=Banner::where('nivel',$nivel)->where('borrado',0)->where('programaestudio_id',$idPrograma)->max('posision');
+        }
         $idban=$queryZero+1;
         return response()->json(["idban"=>$idban]);
-
     }
 }
